+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# master
+Features:
+
+Bug Fixes:
+
+Breaking Changes:
+
 # 3.0.0 rc1
 
 Features: