+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# master
+# 3.0.0 rc2
 Features:
 * Add protocol_version configuration option to allow the user to force the protocol version to use for communication with nodes.
 * Expose listen_address and broadcast_address in `Cassandra::Host` if available.
@@ -6,6 +6,7 @@ Features:
 * Add support for Cassandra indexes in the schema metadata.
 * Add or expose the id, options, keyspace, partition_key, clustering_columns, and clustering_order attributes to table and view schema objects.
 * Add crc_check_chance and extensions attributes to ColumnContainer options.
+* Make cluster configuration options list publicly available. (Thanks, Evan Prothro!)
 
 Bug Fixes:
 * [RUBY-161] Protocol version negotiation in mixed version clusters should not fall back to v1 unless it is truly warranted.